+++ smarts/benchmark/driving_smarts/v2023/metric_formula_platoon.py
@@ -58,15 +58,15 @@ class Formula(FormulaBase):
         """
         params = Params(
             comfort=Comfort(
-                active=False,
-            ),  # TODO: Activate after implementing comfort cost function.
+                active=True,
+            ),
             dist_to_obstacles=DistToObstacles(
                 active=False,
             ),
             gap_between_vehicles=GapBetweenVehicles(
                 active=False,
                 interest="Leader-007",
-            ),  # TODO: Activate after implmenting gap_between_vehicles cost function.
+            ), # TODO: Activate after implementing gap_between_vehicles cost function.
             steps=Steps(
                 active=False,
             ),
@@ -143,7 +143,9 @@ class Formula(FormulaBase):
 
 
 def _humanness(costs: Costs) -> float:
-    humanness = np.array([costs.jerk_linear, costs.lane_center_offset])
+    humanness = np.array(
+        [costs.comfort, costs.lane_center_offset]
+    )
     humanness = np.mean(humanness, dtype=float)
     return 1 - humanness