+++ smarts/env/gymnasium/wrappers/metric/utils.py
@@ -20,11 +20,11 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+from collections import deque
 from dataclasses import fields
 from typing import Callable, TypeVar, Union
 
-from smarts.env.gymnasium.wrappers.metric.costs import Costs
-from smarts.env.gymnasium.wrappers.metric.counts import Counts
+from smarts.env.gymnasium.wrappers.metric.types import Costs, Counts
 
 T = TypeVar("T", Costs, Counts)
 
@@ -95,3 +95,55 @@ def multiply(value: Union[int, float], multiplier: Union[int, float]) -> float:
         float: Value x Multiplier
     """
     return float(value * multiplier)
+
+
+class SlidingWindow:
+    """A sliding window which moves to the right by accepting new elements. The
+    maximum value within the sliding window can be queried at anytime by calling
+    the max() method. 
+    """
+    def __init__(self, size:int):
+        """
+        Args:
+            size (int): Size of the sliding window.
+        """
+        self._values = deque(maxlen=size)
+        self._max_candidates = deque(maxlen=size)
+        self._size = size
+        self._time = -1
+
+    def move(self, x:Union[int,float]):
+        """Moves the sliding window one step to the right by appending the new
+        element x and discarding the oldest element on the left.
+
+        Args:
+            x (Union[int,float]): New element input to the sliding window.
+        """
+        self._time += 1
+
+        # When values deque is full, remove head element of max_candidates deque 
+        # if it matches head element of values deque.
+        if len(self._values) == self._size:
+            if self._values[0][0] == self._max_candidates[0][0]:
+                self._max_candidates.popleft()
+        # Append x to values deque.
+        self._values.append((self._time,x))
+
+	# Remove elements from max_candidates deque's tail which are less than x.
+        while self._max_candidates and self._max_candidates[-1][1] < x:
+            self._max_candidates.pop()
+        # Append x to max_candidates deque.
+        self._max_candidates.append((self._time,x))
+    
+    def max(self):
+        """ Returns the maximum element within the sliding window.
+        """ 
+        return self._max_candidates[0][1]
+    
+    def display(self):
+        """Print the contents of the sliding window.
+        """
+        print("[", end="")
+        for i in self._values:
+            print(i, end=' ')
+        print("]")